requirements.txt
# Core CLI and TUI
typer[all]>=0.9.0
textual>=0.45.0
rich>=13.7.0

# AI and Agents
crewai>=0.28.0
openai>=1.6.0
anthropic>=0.8.0

# Data and Models
pydantic>=2.5.0
sqlalchemy>=2.0.0
alembic>=1.13.0

# Authentication and Security
keyring>=24.3.0
cryptography>=41.0.0
python-dotenv>=1.0.0

# HTTP and APIs
httpx>=0.25.0
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.110.0
msal>=1.25.0

# Utilities
pyyaml>=6.0.1